refactor: decompose oversized functions and split body_model (#126, #127)
Addresses issue #126 by reducing each remaining oversized helper to
<=30 LOC and issue #127 by splitting body_model.py (493 LOC) into
three focused modules each under the 300 LOC module budget.

Issue #126 — decomposed functions:
- drake_models.__main__._build_arg_parser (43 -> 10 LOC):
  extracted _add_exercise_argument, _add_anthropometric_arguments,
  and _add_output_arguments.
- exercises.base.ExerciseModelBuilder._attach_bilateral_grip (44 -> 25 LOC):
  extracted _validate_grip_preconditions (DbC check) and
  _weld_barbell_to_left_hand.
- exercises.base.ExerciseModelBuilder._bilateral_collision_pairs
  (39 -> 13 LOC): extracted _lower_body_collision_pairs and
  _upper_body_collision_pairs.
- exercises.base.ExerciseModelBuilder.build (30 -> 14 LOC):
  extracted _build_body_and_barbell.
- scripts.setup_dev.main (38 -> 15 LOC): extracted
  _require_python_version, _init_submodules, _install_project,
  and _install_vendor_ud_tools.

Issue #127 — split body_model.py:
- body_trunk.py (new, 212 LOC): pelvis, lumbar, torso, head stage
  builders.
- body_limbs.py (new, 228 LOC): upper/lower limb stage builders and
  foot contact geometry.
- body_model.py (493 -> 131 LOC): now a thin orchestrator that only
  owns create_full_body plus back-compatible re-exports of the staged
  helpers and _seg (tests and external callers still import them from
  drake_models.shared.body.body_model).

Pure refactor: no behaviour changes. trajectory_optimizer.py is left
untouched -- issue #118 (DbC contracts) and the #127 entry for the
trajectory optimizer are being addressed in the concurrent issue #142
branch.
